feat(admin): abstrace base_admin class
