+++ typeidea/typeidea/base_admin.py
@@ -1,0 +1,17 @@
+from django.contrib import admin
+
+
+class BaseOwnerAdmin(admin.ModelAdmin):
+    """
+    1.用来自动补充文章、分类、标签、侧边栏、友链这些 Model 的 owner字段
+    2.用来针对 querysite 过滤当前用户数据
+    """
+    exclude = ('owner', )
+
+    def save_model(self, request, obj, form, change):
+        obj.owner = request.user
+        return super(BaseOwnerAdmin, self).save_model(request, obj, form, change)
+
+    def get_queryset(self, request):
+        qs = super(BaseOwnerAdmin, self).get_queryset(request)
+        return qs.filter(owner=request.user)
